backend/requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
pydantic
psycopg2-binary
